% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampd-package.R
\docType{package}
\name{ampd-package}
\alias{ampd}
\alias{ampd-package}
\title{ampd: An Algorithm for Automatic Peak Detection in Noisy Periodic and Quasi-Periodic Signals}
\description{
A method for automatic detection of peaks in noisy periodic and quasi-periodic signals. This method, called automatic multiscale-based peak detection (AMPD), is based on the calculation and analysis of the local maxima scalogram, a matrix comprising the scale-dependent occurrences of local maxima. For further information see \doi{10.3390/a5040588}.
}
\keyword{internal}
